+++ src/components/Inspector/Cover/index.tsx
@@ -1,24 +1,40 @@
 import { Button, DatePicker, DatePickerProps, Input } from "antd";
 import { Controller } from "react-hook-form";
 import styles from "./index.module.scss";
+import { useEffect, useRef } from "react";
 
 export default function CoverInspector() {
-  const postAddressCode = () => {
-    const mapContainer = document.getElementById("map"); // 지도를 표시할 div
-    const mapOption = {
-      center: new window.daum.maps.LatLng(37.537187, 127.005476), // 지도의 중심좌표
-      level: 5, // 지도의 확대 레벨
-    };
+  const map = useRef<any>();
+  const marker = useRef<any>();
 
-    //지도를 미리 생성
-    const map = new window.daum.maps.Map(mapContainer, mapOption);
-    //주소-좌표 변환 객체를 생성
-    const geocoder = new window.daum.maps.services.Geocoder();
-    //마커를 미리 생성
-    const marker = new window.daum.maps.Marker({
-      position: new window.daum.maps.LatLng(37.537187, 127.005476),
-      map: map,
+  useEffect(() => {
+    window.kakao.maps.load(() => {
+      console.log(1234);
+      const mapDiv = document.querySelector("#map");
+      const mapOption = {
+        center: new window.kakao.maps.LatLng(33.450701, 126.570667), // 지도의 중심좌표
+        level: 3, // 지도의 확대 레벨
+      };
+
+      map.current = new window.kakao.maps.Map(mapDiv, mapOption);
+
+      const markerPosition = new window.kakao.maps.LatLng(
+        33.450701,
+        126.570667
+      );
+
+      // 마커를 생성합니다
+      marker.current = new window.kakao.maps.Marker({
+        position: markerPosition,
+      });
+
+      // 마커가 지도 위에 표시되도록 설정합니다
+      marker.current?.setMap(map.current);
     });
+  }, []);
+
+  const postAddressCode = () => {
+    const geocoder = new window.daum.maps.services.Geocoder();
 
     new window.daum.Postcode({
       oncomplete: function (data: any) {
@@ -39,11 +55,11 @@ export default function CoverInspector() {
               // 해당 주소에 대한 좌표를 받아서
               var coords = new window.daum.maps.LatLng(result.y, result.x);
 
-              map.relayout();
+              map.current?.relayout();
               // 지도 중심을 변경한다.
-              map.setCenter(coords);
+              map.current?.setCenter(coords);
               // 마커를 결과값으로 받은 위치로 옮긴다.
-              marker.setPosition(coords);
+              marker.current?.setPosition(coords);
             }
           }
         );
