+++ src/components/Template/A/Map/index.tsx
@@ -10,30 +10,31 @@ interface WedddingMapProps {
 export default function WeddingMap({ onSectionClick }: WedddingMapProps) {
   // https://postcode.map.daum.net/guide
   // https://developers.kakao.com/docs/latest/ko/local/dev-guide#trans-coord
-  useEffect(() => {
-    window.kakao.maps.load(() => {
-      const mapDiv = document.querySelector("#map");
-      const mapOption = {
-        center: new window.kakao.maps.LatLng(33.450701, 126.570667), // 지도의 중심좌표
-        level: 3, // 지도의 확대 레벨
-      };
+  // useEffect(() => {
+  //   window.kakao.maps.load(() => {
+  //     console.log(1234);
+  //     const mapDiv = document.querySelector("#map");
+  //     const mapOption = {
+  //       center: new window.kakao.maps.LatLng(33.450701, 126.570667), // 지도의 중심좌표
+  //       level: 3, // 지도의 확대 레벨
+  //     };
 
-      const map = new window.kakao.maps.Map(mapDiv, mapOption);
+  //     const map = new window.kakao.maps.Map(mapDiv, mapOption);
 
-      const markerPosition = new window.kakao.maps.LatLng(
-        33.450701,
-        126.570667
-      );
+  //     const markerPosition = new window.kakao.maps.LatLng(
+  //       33.450701,
+  //       126.570667
+  //     );
 
-      // 마커를 생성합니다
-      const marker = new window.kakao.maps.Marker({
-        position: markerPosition,
-      });
+  //     // 마커를 생성합니다
+  //     const marker = new window.kakao.maps.Marker({
+  //       position: markerPosition,
+  //     });
 
-      // 마커가 지도 위에 표시되도록 설정합니다
-      marker.setMap(map);
-    });
-  }, []);
+  //     // 마커가 지도 위에 표시되도록 설정합니다
+  //     marker.setMap(map);
+  //   });
+  // }, []);
   return (
     <div className={styles.layout} onClick={onSectionClick}>
       <div className={styles.title}>오시는길</div>
